+++ Source/NetOffice/Availability/SupportedEntityFinder.cs
@@ -5,7 +5,7 @@ using COMTypes = System.Runtime.InteropServices.ComTypes;
 using System.Collections.Generic;
 using NetOffice.Exceptions;
 
-namespace NetOffice.Availity
+namespace NetOffice.Availability
 {
     /// <summary>
     /// Performs a cache supported search to analyze at runtime a proxy supports a method or property
@@ -21,7 +21,7 @@ namespace NetOffice.Availity
         /// <param name="proxy">target com proxy</param>
         /// <param name="name">name of the target entity</param>
         /// <returns>true if supported, otherwise false</returns>
-        /// <exception cref="AvailityException">An unexpected error occurs. See inner exception(s) for details.</exception>
+        /// <exception cref="AvailabilityException">An unexpected error occurs. See inner exception(s) for details.</exception>
         internal bool Find(Core factory, ref Dictionary<string,string> list, SupportedEntityType searchType, object proxy, string name)
         {
             try
@@ -72,8 +72,8 @@ namespace NetOffice.Availity
             }
             catch (Exception exception)
             {
-                throw new AvailityException(exception);
-            }            
+                throw new AvailabilityException(exception);
+            }
         }
 
         /// <summary>
@@ -86,7 +86,7 @@ namespace NetOffice.Availity
         internal Dictionary<string, string> GetSupportedEntities(Core factory, object comProxy)
         {
             try
-            {                
+            {
                 Guid parentLibraryGuid = CoreFactoryExtensions.GetParentLibraryGuid(factory, comProxy);
                 if (Guid.Empty == parentLibraryGuid)
                     return null;